Added support for local LLM llama and vision to generate description
