+++ LICENSE
@@ -1,6 +1,7 @@
 MIT License
 
 Copyright (c) 2024 Matt Adams
+Copyright (c) 2024 Shane Tan
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
