+++ helper_functions.py
@@ -1,0 +1,501 @@
+import json
+import re
+import streamlit as st
+import streamlit.components.v1 as components
+from mistralai.client import MistralClient
+from mistralai.models.chat_completion import ChatMessage
+from openai import OpenAI
+from openai import AzureOpenAI
+import ollama
+from PIL import Image
+import base64
+
+# ------------------ Helper Functions ------------------ #
+
+# Function to encode the image
+def encode_image(image_file):
+    b64_image = base64.b64encode(image_file).decode()
+    # return b64_image
+    return f"data:image/png;base64,{b64_image}"
+
+# Function to get user image upload
+def get_image_input():
+    uploaded_file = st.file_uploader("Choose an image...", type=["jpg", "png"])
+
+    if uploaded_file is not None:
+        image = Image.open(uploaded_file)
+        st.image(image, caption='Uploaded Image.', width=200)
+        return uploaded_file.getvalue()
+        # image_base64 = encode_image(uploaded_file.getvalue())
+        # return image_base64
+    
+# Function to create a prompt for generating a threat model
+def create_image_description_prompt(image_base64):
+    prompt = f"""
+Act as a cyber security expert with more than 20 years experience of using the STRIDE threat modelling methodology to produce comprehensive threat models for a wide range of applications. Your task is to describe the image in detail for the purpose of threat modeling.
+
+With the following image, identify components, the role of the component, and the connections. Include any other relevant information that can be useful to identify potential threats.
+
+![Uploaded Image](data:image/png;base64,{image_base64})
+"""
+    return prompt
+
+# Function to get image description from the GPT response.
+def get_image_description(api_key, model_name, prompt):
+    client = OpenAI(api_key=api_key)
+
+    response = client.ChatCompletion.create(
+        model=model_name,
+        messages=[
+            {"role": "system", "content": "You are a helpful assistant designed to output JSON."},
+            {"role": "user", "content": prompt}
+        ],
+        max_tokens=4000,
+    )
+
+    # Convert the JSON string in the 'content' field to a Python dictionary
+    response_content = json.loads(response.choices[0].message.content)
+
+    return response_content
+
+# Function to get image description from lcocal model
+def get_image_description_local(model_name, prompt):
+    response = ollama.chat(
+        model=model_name,
+        format='json',
+        messages=[
+            {"role": "system", "content": "You are a helpful assistant designed to output JSON."},
+            {"role": "user", "content": prompt}
+        ],
+    )
+
+    # Convert the JSON string in the 'content' field to a Python dictionary
+    response_content = json.loads(response['message']['content'])
+
+    return response_content
+
+# Function to get user input for the application description and key details
+def get_input():
+
+    input_text = st.text_area(
+        label="Describe the application to be modelled",
+        placeholder="Enter your application details...",
+        value=f"""A web application that allows users to create, store, and share personal notes.
+        Application consist of a web and database. users access system via internet over http. web server connects to database server using mongodb port 27017 without encryption. 
+        The application is built using the React frontend framework and a Node.js backend with a MongoDB database. Users can sign up for an account and log in using OAuth2 with Google or Facebook. The notes are encrypted at rest and are only accessible by the user who created them. The application also supports real-time collaboration on notes with other users.
+        Logging is not setup on all the servers. There is no firewall protecting the web application.""",
+        height=150,
+        key="app_input",
+        help="Please provide a detailed description of the application, including the purpose of the application, the technologies used, and any other relevant information.",
+    )
+    return input_text
+
+# Function to create a prompt for generating a threat model
+def create_threat_model_prompt(app_type, authentication, internet_facing, sensitive_data, pam, app_input):
+    prompt = f"""
+Act as a cyber security expert with more than 20 years experience of using the STRIDE threat modelling methodology to produce comprehensive threat models for a wide range of applications. Your task is to use the application description and additional provided to you to produce a list of specific threats for the application.
+
+For each of the STRIDE categories (Spoofing, Tampering, Repudiation, Information Disclosure, Denial of Service, and Elevation of Privilege), list multiple (3 or 4) credible threats if applicable. Each threat scenario should provide a credible scenario in which the threat could occur in the context of the application. It is very important that your responses are tailored to reflect the details you are given.
+
+
+When providing the threat model, use a JSON formatted response with the keys "threat_model" and "improvement_suggestions". Under "threat_model", include an array of objects with the keys "Threat Type", "Scenario", and "Potential Impact". 
+
+Under "improvement_suggestions", include an array of strings with suggestions on how the threat modeller can improve their application description in order to allow the tool to produce a more comprehensive threat model.
+
+APPLICATION TYPE: {app_type}
+AUTHENTICATION METHODS: {authentication}
+INTERNET FACING: {internet_facing}
+SENSITIVE DATA: {sensitive_data}
+PRIVILEGED ACCESS MANAGEMENT: {pam}
+APPLICATION DESCRIPTION: {app_input}
+
+Example of expected JSON response format:
+  
+    {{
+      "threat_model": [
+        {{
+          "Threat Type": "Spoofing",
+          "Scenario": "Example Scenario 1",
+          "Potential Impact": "Example Potential Impact 1"
+        }},
+        {{
+          "Threat Type": "Spoofing",
+          "Scenario": "Example Scenario 2",
+          "Potential Impact": "Example Potential Impact 2"
+        }},
+        // ... more threats
+      ],
+      "improvement_suggestions": [
+        "Example improvement suggestion 1.",
+        "Example improvement suggestion 2.",
+        // ... more suggestions
+      ]
+    }}
+"""
+    return prompt
+
+# Function to get threat model from the GPT response.
+def get_threat_model(api_key, model_name, prompt):
+    client = OpenAI(api_key=api_key)
+
+    response = client.chat.completions.create(
+        model=model_name,
+        response_format={"type": "json_object"},
+        messages=[
+            {"role": "system", "content": "You are a helpful assistant designed to output JSON."},
+            {"role": "user", "content": prompt}
+        ],
+        max_tokens=4000,
+    )
+
+    # Convert the JSON string in the 'content' field to a Python dictionary
+    response_content = json.loads(response.choices[0].message.content)
+
+    return response_content
+
+# Function to get threat model from the Azure OpenAI response.
+def get_threat_model_azure(api_endpoint, api_key, api_version, deployment_name, prompt):
+    client = AzureOpenAI(
+        azure_endpoint = api_endpoint,
+        api_key = api_key,
+        api_version = api_version,
+    )
+
+    response = client.chat.completions.create(
+        model = deployment_name,
+        response_format={"type": "json_object"},
+        messages=[
+            {"role": "system", "content": "You are a helpful assistant designed to output JSON."},
+            {"role": "user", "content": prompt}
+        ]
+    )
+
+    # Convert the JSON string in the 'content' field to a Python dictionary
+    response_content = json.loads(response.choices[0].message.content)
+
+    return response_content
+
+# Function to get threat model from the Azure OpenAI response.
+def get_threat_model_mistral(api_key, model_name, prompt):
+    client = MistralClient(api_key=api_key)
+
+    response = client.chat(
+        model = model_name,
+        response_format={"type": "json_object"},
+        messages=[
+            ChatMessage(role="user", content=prompt)
+        ]
+    )
+
+    # Convert the JSON string in the 'content' field to a Python dictionary
+    response_content = json.loads(response.choices[0].message.content)
+
+    return response_content
+
+# Function to get threat model from the GPT response.
+def get_threat_model_local(model_name, prompt):
+    response = ollama.chat(
+        model=model_name,
+        format='json',
+        messages=[
+            {"role": "system", "content": "You are a helpful assistant designed to output JSON."},
+            {"role": "user", "content": prompt}
+        ],
+    )
+
+    # Convert the JSON string in the 'content' field to a Python dictionary
+    response_content = json.loads(response['message']['content'])
+
+    return response_content
+
+# Function to convert JSON to Markdown for display.    
+def json_to_markdown(threat_model, improvement_suggestions):
+    markdown_output = "## Threat Model\n\n"
+    
+    # Start the markdown table with headers
+    markdown_output += "| Threat Type | Scenario | Potential Impact |\n"
+    markdown_output += "|-------------|----------|------------------|\n"
+    
+    # Fill the table rows with the threat model data
+    for threat in threat_model:
+        markdown_output += f"| {threat['Threat Type']} | {threat['Scenario']} | {threat['Potential Impact']} |\n"
+    
+    markdown_output += "\n\n## Improvement Suggestions\n\n"
+    for suggestion in improvement_suggestions:
+        markdown_output += f"- {suggestion}\n"
+    
+    return markdown_output
+
+
+# Function to create a prompt to generate an attack tree
+def create_attack_tree_prompt(app_type, authentication, internet_facing, sensitive_data, pam, app_input):
+    prompt = f"""
+APPLICATION TYPE: {app_type}
+AUTHENTICATION METHODS: {authentication}
+INTERNET FACING: {internet_facing}
+SENSITIVE DATA: {sensitive_data}
+PRIVILEGED ACCESS MANAGEMENT: {pam}
+APPLICATION DESCRIPTION: {app_input}
+"""
+    return prompt
+
+
+# Function to get attack tree from the GPT response.
+def get_attack_tree(api_key, model_name, prompt):
+    client = OpenAI(api_key=api_key)
+
+    response = client.chat.completions.create(
+        model=model_name,
+        messages=[
+            {"role": "system", "content": """
+Act as a cyber security expert with more than 20 years experience of using the STRIDE threat modelling methodology to produce comprehensive threat models for a wide range of applications. Your task is to use the application description provided to you to produce an attack tree in Mermaid syntax. The attack tree should reflect the potential threats for the application based on the details given.
+
+You MUST only respond with the Mermaid code block. See below for a simple example of the required format and syntax for your output.
+
+```mermaid
+graph TD
+    A[Enter Chart Definition] --> B(Preview)
+    B --> C{{decide}}
+    C --> D["Keep"]
+    C --> E["Edit Definition (Edit)"]
+    E --> B
+    D --> F["Save Image and Code"]
+    F --> B
+```
+
+IMPORTANT: Round brackets are special characters in Mermaid syntax. If you want to use round brackets inside a node label you MUST wrap the label in double quotes. For example, ["Example Node Label (ENL)"].
+"""},
+            {"role": "user", "content": prompt}
+        ]
+    )
+
+    # Access the 'content' attribute of the 'message' object directly
+    attack_tree_code = response.choices[0].message.content
+    
+    # Remove Markdown code block delimiters using regular expression
+    attack_tree_code = re.sub(r'^```mermaid\s*|\s*```$', '', attack_tree_code, flags=re.MULTILINE)
+
+    return attack_tree_code
+
+# Function to get attack tree from the Azure OpenAI response.
+def get_attack_tree_azure(api_endpoint, api_key, api_version, deployment_name, prompt):
+    client = AzureOpenAI(
+        azure_endpoint = api_endpoint,
+        api_key = api_key,
+        api_version = api_version,
+    )
+
+    response = client.chat.completions.create(
+        model = deployment_name,
+        messages=[
+            {"role": "system", "content": """
+Act as a cyber security expert with more than 20 years experience of using the STRIDE threat modelling methodology to produce comprehensive threat models for a wide range of applications. Your task is to use the application description provided to you to produce an attack tree in Mermaid syntax. The attack tree should reflect the potential threats for the application based on the details given.
+
+You MUST only respond with the Mermaid code block. See below for a simple example of the required format and syntax for your output.
+
+```mermaid
+graph TD
+    A[Enter Chart Definition] --> B(Preview)
+    B --> C{{decide}}
+    C --> D["Keep"]
+    C --> E["Edit Definition (Edit)"]
+    E --> B
+    D --> F["Save Image and Code"]
+    F --> B
+```
+
+IMPORTANT: Round brackets are special characters in Mermaid syntax. If you want to use round brackets inside a node label you MUST wrap the label in double quotes. For example, ["Example Node Label (ENL)"].
+"""},
+            {"role": "user", "content": prompt}
+        ]
+    )
+
+    # Access the 'content' attribute of the 'message' object directly
+    attack_tree_code = response.choices[0].message.content
+    
+    # Remove Markdown code block delimiters using regular expression
+    attack_tree_code = re.sub(r'^```mermaid\s*|\s*```$', '', attack_tree_code, flags=re.MULTILINE)
+
+    return attack_tree_code
+
+
+# Function to get attack tree from the Mistral model's response.
+def get_attack_tree_mistral(api_key, model_name, prompt):
+    client = MistralClient(api_key=api_key)
+
+    response = client.chat(
+        model=model_name,
+        messages=[
+            {"role": "system", "content": """
+Act as a cyber security expert with more than 20 years experience of using the STRIDE threat modelling methodology to produce comprehensive threat models for a wide range of applications. Your task is to use the application description provided to you to produce an attack tree in Mermaid syntax. The attack tree should reflect the potential threats for the application based on the details given.
+
+You MUST only respond with the Mermaid code block. See below for a simple example of the required format and syntax for your output.
+
+```mermaid
+graph TD
+    A[Enter Chart Definition] --> B(Preview)
+    B --> C{{decide}}
+    C --> D["Keep"]
+    C --> E["Edit Definition (Edit)"]
+    E --> B
+    D --> F["Save Image and Code"]
+    F --> B
+```
+
+IMPORTANT: Round brackets are special characters in Mermaid syntax. If you want to use round brackets inside a node label you MUST wrap the label in double quotes. For example, ["Example Node Label (ENL)"].
+"""},
+            {"role": "user", "content": prompt}
+        ]
+    )
+
+    # Access the 'content' attribute of the 'message' object directly
+    attack_tree_code = response.choices[0].message.content
+    
+    # Remove Markdown code block delimiters using regular expression
+    attack_tree_code = re.sub(r'^```mermaid\s*|\s*```$', '', attack_tree_code, flags=re.MULTILINE)
+
+    return attack_tree_code
+
+# Function to get attack tree from the Local model's response.
+def get_attack_tree_local(model_name, prompt):
+
+    response = ollama.chat(
+        model=model_name,
+        messages=[
+            {"role": "system", "content": """
+Act as a cyber security expert with more than 20 years experience of using the STRIDE threat modelling methodology to produce comprehensive threat models for a wide range of applications. Your task is to use the application description provided to you to produce an attack tree in Mermaid syntax. The attack tree should reflect the potential threats for the application based on the details given.
+
+You MUST only respond with the Mermaid code block. See below for a simple example of the required format and syntax for your output.
+
+```mermaid
+graph TD
+    A[Enter Chart Definition] --> B(Preview)
+    B --> C{{decide}}
+    C --> D["Keep"]
+    C --> E["Edit Definition (Edit)"]
+    E --> B
+    D --> F["Save Image and Code"]
+    F --> B
+```
+
+IMPORTANT: Round brackets are special characters in Mermaid syntax. If you want to use round brackets inside a node label you MUST wrap the label in double quotes. For example, ["Example Node Label (ENL)"].
+"""},
+            {"role": "user", "content": prompt}
+        ]
+    )
+
+    # Access the 'content' attribute of the 'message' object directly
+    attack_tree_code = response['message']['content']
+    
+    # Remove Markdown code block delimiters using regular expression
+    attack_tree_code = re.sub(r'^```mermaid\s*|\s*```$', '', attack_tree_code, flags=re.MULTILINE)
+
+    return attack_tree_code
+
+# Function to render Mermaid diagram
+def mermaid(code: str, height: int = 500) -> None:
+    components.html(
+        f"""
+        <pre class="mermaid" style="height: {height}px;">
+            {code}
+        </pre>
+
+        <script type="module">
+            import mermaid from 'https://cdn.jsdelivr.net/npm/mermaid@10/dist/mermaid.esm.min.mjs';
+            mermaid.initialize({{ startOnLoad: true }});
+        </script>
+        """,
+        height=height,
+    )
+
+
+# Function to create a prompt to generate mitigating controls
+def create_mitigations_prompt(threats):
+    prompt = f"""
+Act as a cyber security expert with more than 20 years experience of using the STRIDE threat modelling methodology. Your task is to provide potential mitigations for the threats identified in the threat model. It is very important that your responses are tailored to reflect the details of the threats.
+
+Your output should be in the form of a markdown table with the following columns:
+    - Column A: Threat Type
+    - Column B: Scenario
+    - Column C: Suggested Mitigation(s)
+
+Below is the list of identified threats:
+{threats}
+
+YOUR RESPONSE (do not wrap in a code block):
+"""
+    return prompt
+
+
+# Function to get mitigations from the GPT response.
+def get_mitigations(api_key, model_name, prompt):
+    client = OpenAI(api_key=api_key)
+
+    response = client.chat.completions.create(
+        model = model_name,
+        messages=[
+            {"role": "system", "content": "You are a helpful assistant that provides threat mitigation strategies in Markdown format."},
+            {"role": "user", "content": prompt}
+        ]
+    )
+
+    # Access the content directly as the response will be in text format
+    mitigations = response.choices[0].message.content
+
+    return mitigations
+
+
+# Function to get mitigations from the Azure OpenAI response.
+def get_mitigations_azure(api_endpoint, api_key, api_version, deployment_name, prompt):
+    client = AzureOpenAI(
+        azure_endpoint = api_endpoint,
+        api_key = api_key,
+        api_version = api_version,
+    )
+
+    response = client.chat.completions.create(
+        model = deployment_name,
+        messages=[
+            {"role": "system", "content": "You are a helpful assistant that provides threat mitigation strategies in Markdown format."},
+            {"role": "user", "content": prompt}
+        ]
+    )
+
+    # Access the content directly as the response will be in text format
+    mitigations = response.choices[0].message.content
+
+    return mitigations
+
+
+# Function to get mitigations from the Mistral model's response.
+def get_mitigations_mistral(api_key, model_name, prompt):
+    client = MistralClient(api_key=api_key)
+
+    response = client.chat(
+        model = model_name,
+        messages=[
+            {"role": "system", "content": "You are a helpful assistant that provides threat mitigation strategies in Markdown format."},
+            {"role": "user", "content": prompt}
+        ]
+    )
+
+    # Access the content directly as the response will be in text format
+    mitigations = response.choices[0].message.content
+
+    return mitigations
+
+# Function to get mitigations from the local model's response.
+def get_mitigations_local(model_name, prompt):
+
+    response = ollama.chat(
+        model = model_name,
+        messages=[
+            {"role": "system", "content": "You are a helpful assistant that provides threat mitigation strategies in Markdown format."},
+            {"role": "user", "content": prompt}
+        ]
+    )
+
+    # Access the content directly as the response will be in text format
+    mitigations = response['message']['content']
+
+    return mitigations
